Extract image class from decoder class
